refactor(components): break out file preview to separate components
