+++ apps/www/content/docs/components/message-input.mdx
@@ -48,6 +48,10 @@ npx shadcn@latest add https://shadcn-chatbot-kit.vercel.app/r/message-input.json
 npm install framer-motion@11 remeda@2
 ```
 
+<Step>Add the `FilePreview` component to your project.</Step>
+
+The `MessageInput` component uses the `FilePreview` component. Make sure you have it installed in your project.
+
 <Step>Copy and paste the following code into your project.</Step>
 
 `components/ui/message-input.tsx`
